CHORE :: Detail -> Result Rename
