+++ components/ChatResult/ChatResultInput/style.ts
