+++ components/ChatResult/ChatResultRecipe/style.ts
