+++ components/ChatResult/ChatResultSection/index.tsx
@@ -4,10 +4,10 @@ import * as CS from "@/components/Chat/ChatSection/style";
 import { useRecoilValue } from "recoil";
 import { MaterialListState } from "@/atoms/Chat/MaterialList";
 import ChatInput from "@/components/Chat/ChatInput";
-import ChatDetailInput from "../ChatDetailInput";
-import ChatDetailRecipe from "../ChatDetailRecipe";
+import ChatResultInput from "../ChatResultInput";
+import ChatResultRecipe from "../ChatResultRecipe";
 
-function ChatDetailSection() {
+function ChatResultSection() {
   const materialList = useRecoilValue(MaterialListState);
   return (
     <>
@@ -28,26 +28,26 @@ function ChatDetailSection() {
               <CS.ContentLabel>레시피 :</CS.ContentLabel>
             </CS.Content>
             <S.RecipeList>
-              <ChatDetailRecipe index={1} title="김치 감자 피망 돼지고기 볶음">
+              <ChatResultRecipe index={1} title="김치 감자 피망 돼지고기 볶음">
                 감자와 피망을 굵게 썬 후 돼지고기와 함께 볶은 뒤 김치를 넣고
                 볶아줍니다.
-              </ChatDetailRecipe>
-              <ChatDetailRecipe index={2} title="돼지고기 김치 찌개">
+              </ChatResultRecipe>
+              <ChatResultRecipe index={2} title="돼지고기 김치 찌개">
                 돼지고기와 김치를 함께 끓여 매운 김치찌개를 만듭니다. 감자와
                 피망을 넣어 더욱 맛을 높일 수 있습니다.
-              </ChatDetailRecipe>
-              <ChatDetailRecipe index={3} title="김치 감자 피망 돼지고기 전">
+              </ChatResultRecipe>
+              <ChatResultRecipe index={3} title="김치 감자 피망 돼지고기 전">
                 감자와 피망, 김치를 다져 돼지고기와 함께 새콤한 전을 만들어
                 먹습니다.
-              </ChatDetailRecipe>
+              </ChatResultRecipe>
             </S.RecipeList>
             <CS.Content>마음에 드시는 요리를 선택하세요!</CS.Content>
           </CS.ContentWrap>
         </CS.ChatWrap>
-        <ChatDetailInput />
+        <ChatResultInput />
       </CS.ChatSection>
     </>
   );
 }
 
-export default ChatDetailSection;
+export default ChatResultSection;
