+++ components/ChatResult/ChatResultSection/style.ts
