+++ components/ChatResult/style.ts
